requirements.txt
urwid
pyyaml
jinja2
requests_oauthlib
ws4py==0.3.4
nose
tox
bson
termcolor
petname
six
configobj
progressbar2
prettytable
q
